extend domVisitor to support multiple audit logics
